done-delete-basic-version

+++ assets/src/send-data.js
@@ -11,13 +11,13 @@ class Ulubione {
 		document.addEventListener("DOMContentLoaded", () => {
 			document.querySelectorAll(".ulub").forEach((button) => {
 				button.addEventListener("click", (e) =>
-					this.clickUlubhendler(e)
+					this.clickUlubheHandller(e)
 				);
 			});
 		});
 	}
 
-	async clickUlubhendler(e) {
+	async clickUlubheHandller(e) {
 		e.preventDefault();
 
 		let currentClicked = e.target.closest("button");
@@ -41,35 +41,32 @@ class Ulubione {
 					),
 				}
 			);
-			currentClicked.setAttribute("");
 		} catch {
-			console.error("Coś poszło Mucho Grande nie tak");
+			console.error();
 		}
 	}
 
 	async deletePost(currentClicked) {
-		const favId = currentClicked.getAttribute("");
-		if (favId) {
-			currentClicked.setAttribute("data-ulubione-", favId);
-			console.log("Udało się wygenerować unikalne ID");
-		} else {
-			console.log("Nie wygenerowano nowegoID");
+		const originalID = Number(
+			currentClicked.getAttribute("data-original-post-id")
+		);
+		if (!Number.isFinite(originalID)) {
+			console.warn(
+				"Nie udało się ustawić akceptowalnej wartości, atrybut data musi być liczbą"
+			);
+			return;
 		}
 		try {
 			const response = await axios.delete(
 				localAddress + "/wp-json/ulub/v1/sendUlub",
 				{
 					data: {
-						list_type:
-							currentClicked.getAttribute("data-list-type"),
-						favorite_post_id:
-							currentClicked.getAttribute("favorite_post_id"),
+						original_post_id: originalID,
 					},
 				}
 			);
-		} catch (error) {
-			console.error(error);
-			console.warn("Brak data-ulubione-id – nie wiem co usunąć");
+		} catch {
+			console.error();
 		}
 	}
 }
